Project Started; Great Progress All Around
Project three has been started, and I'm much more confident on how this one is going so far in terms of the JavaScript code. There still is much left to do however, but it luckily isn't too much, and with documentation also done along the way, I'm sure I'll be getting this down by tomorrow.