Project Completed; Ready for Review
There are a lot of errors in my syntax of code, but getting through all the steps in the steps in the book is all I'm willing to do as I just don't find a way to solve many of the obtuse questions to work as it should. At least the documentation is correct to what it should be doing.